requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements.txt requirements.in
#
asgiref==3.7.2
    # via django
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20230430
    # via django-import-export
django==4.1.10
    # via
    #   -r requirements.in
    #   django-import-export
    #   django-leaflet
django-import-export==3.2.0
    # via -r requirements.in
django-leaflet==0.29.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
markuppy==1.14
    # via tablib
marshmallow==3.19.0
    # via -r requirements.in
numpy==1.25.1
    # via pandas
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
packaging==23.1
    # via marshmallow
pandas==2.0.3
    # via -r requirements.in
psycopg2==2.9.6
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0
    # via tablib
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.5.0
    # via django-import-export
tzdata==2023.3
    # via pandas
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
